# api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PudimFinance API",
    "description": "Personal finance application API. Immutable double-entry ledger, event sourcing, real-time reporting, receipt scanning and price tracking.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/audit/events": {
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "Lists audit events for administrators.",
        "operationId": "list_audit_events",
        "responses": {
          "200": {
            "description": "List of audit events"
          },
          "403": {
            "description": "Admin access required"
          }
        }
      }
    },
...
  },
...
  "security": [
    {}
  ],
...
}